requirements.txt
Django==3.0.7
Django==3.0.7
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pygments==2.6.1
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework==3.11.0
idna==2.9
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
psycopg2
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
